add confirmation on delete

+++ GamesPlay/client/src/components/game-details/action-btns/ActionBtns.jsx
@@ -1,11 +1,16 @@
 import { Link, useNavigate } from 'react-router-dom';
 import { useDeleteGame } from '../../../hooks/useGames.js';
 
-export default function ActionBtns({ gameId }) {
+export default function ActionBtns({ gameId, title }) {
     const navigate = useNavigate();
     const deleteGame = useDeleteGame();
 
     async function deleteGameHandler() {
+        const isConfirmed = confirm(`Are you sure you want to delete ${title} game?`);
+        if (!isConfirmed) {
+            return;
+        }
+
         try {
             await deleteGame(gameId);
             navigate('/');
